orchestrator/requirements.txt
fastapi
uvicorn[standard]
kafka-python
SQLAlchemy